# openapi/mygraph.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SAP B1 Knowledge Graph API",
    "description": "Read access to the BusinessPartner + Invoice knowledge graph ingested from SAP B1 Service Layer.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://mygraph.cfapps.us10-001.hana.ondemand.com"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Liveness check",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/graph/bp/{cardCode}": {
      "get": {
        "operationId": "getBusinessPartnerGraph",
        "summary": "Get a BusinessPartner's invoice summary (default) or full invoice/line detail",
        "description": "By default returns a compact summary: core BusinessPartner fields, AR/AP invoice counts and totals, and the 5 most recent invoices (headers only, no line items). This keeps the response well under the size limits of consumers like SAP Joule Studio Actions, which log the full response and reject oversized payloads. Pass ?summary=false for the full subgraph (every invoice with every line) -- only use that for direct API/agent consumers without a response-size limit, not for the Joule Action.",
        "parameters": [
          {
            "name": "cardCode",
            "in": "path",
            "required": true,
            "description": "SAP B1 BusinessPartner CardCode",
            "schema": {
              "type": "string"
            },
            "example": "C20000"
          },
          {
            "name": "summary",
            "in": "query",
            "required": false,
            "description": "Set to \"false\" to get the full subgraph (every invoice + every line) instead of the summary. Defaults to true.",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "default": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "BusinessPartner summary (default) or full subgraph (?summary=false)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessPartnerSummary"
                }
              }
            }
          },
          "404": {
            "description": "No BusinessPartner with this CardCode in the graph",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Graph store has not been populated yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BusinessPartnerProperties": {
        "type": "object",
        "properties": {
          "CardCode": {
            "type": "string"
          },
          "CardName": {
            "type": "string",
            "nullable": true
          },
          "CardType": {
            "type": "string",
            "enum": [
              "cCustomer",
              "cSupplier",
              "cLid"
            ]
          },
          "GroupCode": {
            "type": "integer",
            "nullable": true
          },
          "Currency": {
            "type": "string",
            "nullable": true
          },
          "Valid": {
            "type": "string",
            "enum": [
              "tYES",
              "tNO"
            ],
            "nullable": true
          },
          "Frozen": {
            "type": "string",
            "enum": [
              "tYES",
              "tNO"
            ],
            "nullable": true
          },
          "Balance": {
            "type": "number",
            "nullable": true
          },
          "CreditLimit": {
            "type": "number",
            "nullable": true
          }
        },
        "additionalProperties": true
      },
      "InvoiceProperties": {
        "type": "object",
        "properties": {
          "DocEntry": {
            "type": "integer"
          },
          "DocNum": {
            "type": "integer",
            "nullable": true
          },
          "CardCode": {
            "type": "string"
          },
          "CardName": {
            "type": "string",
            "nullable": true
          },
          "DocDate": {
            "type": "string",
            "nullable": true
          },
          "DocDueDate": {
            "type": "string",
            "nullable": true
          },
          "DocTotal": {
            "type": "number"
          },
          "DocCurrency": {
            "type": "string",
            "nullable": true
          },
          "DocumentStatus": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "type": "string",
            "enum": [
              "AR",
              "AP"
            ]
          }
        },
        "additionalProperties": false
      },
      "InvoiceLineProperties": {
        "type": "object",
        "properties": {
          "LineNum": {
            "type": "integer"
          },
          "ItemCode": {
            "type": "string",
            "nullable": true
          },
          "ItemDescription": {
            "type": "string",
            "nullable": true
          },
          "Quantity": {
            "type": "number",
            "nullable": true
          },
          "Price": {
            "type": "number",
            "nullable": true
          },
          "LineTotal": {
            "type": "number",
            "nullable": true
          },
          "WarehouseCode": {
            "type": "string",
            "nullable": true
          },
          "AccountCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BusinessPartnerNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "BP:C20000"
          },
          "type": {
            "type": "string",
            "enum": [
              "BusinessPartner"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/BusinessPartnerProperties"
          }
        },
        "additionalProperties": false
      },
      "InvoiceNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "Invoice:AR:1"
          },
          "type": {
            "type": "string",
            "enum": [
              "Invoice"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/InvoiceProperties"
          }
        },
        "additionalProperties": false
      },
      "InvoiceLineNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "Invoice:AR:1:Line:0"
          },
          "type": {
            "type": "string",
            "enum": [
              "InvoiceLine"
            ]
          },
          "properties": {
            "$ref": "#/components/schemas/InvoiceLineProperties"
          }
        },
        "additionalProperties": false
      },
      "InvoiceWithLines": {
        "type": "object",
        "properties": {
          "invoice": {
            "$ref": "#/components/schemas/InvoiceNode"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineNode"
            }
          }
        },
        "additionalProperties": false
      },
      "InvoiceCountAndTotal": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "totalDocTotal": {
            "type": "number"
          }
        },
        "additionalProperties": false
      },
      "RecentInvoiceSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "Invoice:AR:609"
          },
          "direction": {
            "type": "string",
            "enum": [
              "AR",
              "AP"
            ]
          },
          "docNum": {
            "type": "integer",
            "nullable": true
          },
          "docDate": {
            "type": "string",
            "nullable": true
          },
          "docTotal": {
            "type": "number"
          },
          "documentStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BusinessPartnerSummary": {
        "type": "object",
        "description": "Default response shape for GET /graph/bp/{cardCode} -- compact, safe for size-limited consumers.",
        "properties": {
          "businessPartner": {
            "type": "object",
            "properties": {
              "CardCode": {
                "type": "string"
              },
              "CardName": {
                "type": "string",
                "nullable": true
              },
              "CardType": {
                "type": "string",
                "nullable": true
              },
              "Currency": {
                "type": "string",
                "nullable": true
              },
              "Balance": {
                "type": "number",
                "nullable": true
              },
              "CreditLimit": {
                "type": "number",
                "nullable": true
              }
            },
            "additionalProperties": false
          },
          "invoiceSummary": {
            "type": "object",
            "properties": {
              "AR": {
                "$ref": "#/components/schemas/InvoiceCountAndTotal"
              },
              "AP": {
                "$ref": "#/components/schemas/InvoiceCountAndTotal"
              }
            },
            "additionalProperties": false
          },
          "recentInvoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentInvoiceSummary"
            }
          }
        },
        "additionalProperties": false
      },
      "BusinessPartnerSubgraph": {
        "type": "object",
        "properties": {
          "businessPartner": {
            "$ref": "#/components/schemas/BusinessPartnerNode"
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceWithLines"
            }
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}